Move wifi setup commands into separate file
