+++ setup.py
@@ -71,8 +71,8 @@ setup(
         [paste.app_factory]
         main = senic_hub.backend:main
         [console_scripts]
-        wifi_setup = senic_hub.backend.commands:wifi_setup
-        scan_wifi = senic_hub.backend.commands:scan_wifi
+        wifi_setup = senic_hub.backend.wifi_setup:wifi_setup
+        scan_wifi = senic_hub.backend.wifi_setup:scan_wifi
         create_configurations = senic_hub.backend.commands:create_configuration_files_and_restart_apps
         device_discovery = senic_hub.backend.commands:device_discovery
         nuimo_app = senic_hub.nuimo_app.__main__:main
